+++ src/XamU.Core/Interfaces/INavigationService.cs
@@ -61,8 +61,28 @@ namespace XamarinUniversity.Infrastructure
         /// </summary>
         /// <returns>The async.</returns>
         /// <param name="key">Navigation key.</param>
-        /// <param name="state">State (often View model)</param>
-        Task NavigateAsync(object key, object state = null);
+        Task NavigateAsync(object key);
+
+        /// <summary>
+        /// Navigate to a page using the known key.
+        /// The state object (if not null) will be assigned as the BindingContext
+        /// if the View doesn't assign it as part of construction. If there _is_ a BindingContext,
+        /// then the state will be passed to the IViewModelNavigationInit.IntializeAsync implementation.
+        /// </summary>
+        /// <returns>Task</returns>
+        /// <param name="key">Navigation key.</param>
+        /// <param name="state">State (can be ViewModel)</param>
+        Task NavigateAsync(object key, object state);
+
+        /// <summary>
+        /// Navigate to a page using the known key. Assigns the passed ViewModel as the
+        /// BindingContext, and supplies the state to the IViewModelNavigationInit.IntializeAsync implementation.
+        /// </summary>
+        /// <returns>Task</returns>
+        /// <param name="key">Navigation key.</param>
+        /// <param name="viewModel">ViewModel</param>
+        /// <param name="state">Optional state parameter</param>
+        Task NavigateAsync(object key, object viewModel, object state);
 
         /// <summary>
         /// Returns true/false whether we can go backwards on the Nav Stack.
@@ -81,8 +101,24 @@ namespace XamarinUniversity.Infrastructure
         /// </summary>
         /// <returns>Async response</returns>
         /// <param name="key">Navigation key.</param>
+        Task PushModalAsync(object key);
+
+        /// <summary>
+        /// Push a page onto the modal stack.
+        /// </summary>
+        /// <returns>Async response</returns>
+        /// <param name="key">Navigation key.</param>
         /// <param name="state">State (often ViewModel)</param>
-        Task PushModalAsync(object key, object state = null);
+        Task PushModalAsync(object key, object state);
+
+        /// <summary>
+        /// Push a page onto the modal stack.
+        /// </summary>
+        /// <returns>Async response</returns>
+        /// <param name="key">Navigation key.</param>
+        /// <param name="viewModel">ViewModel to assign (can be null)</param>
+        /// <param name="state">Initialization state (can be null)</param>
+        Task PushModalAsync(object key, object viewModel, object state);
 
         /// <summary>
         /// Pops the last page off the modal stack